[Basic] Rename `abortWithPrettyStackTraceMessage` to `ABORT`
Turn it into a wrapping macro that includes the file location, and
move into `Assertions.h`.
